* master: (27 commits)
  [BUA-864] Changing how tests are executed, so travis only retries for the integration that failed.
  [BUA-864] Following SQ advices.
  [BUA-864] Using global variable to check for the container log.
  [BUA-864] Fixing and/or removing tests, updating badge link.
  [BUA-864] Grabbing docker-selenium capabilities from the latest image present.
  [BUA-864] Refactoring some duplicated code.
  [BUA-864] Declaring explicit scope for methods.
  [BUA-864] Fixing explicit scope for methods.
  [BUA-864] Using explicit scope for methods.
  [BUA-864] Using explicit scope for methods.
  [BUA-864] Updating readme.
  [BUA-864] Using exported main VNC port to build live preview URL.
  [BUA-864] Using exported container name to start Zalenium with any container name.
  [BUA-864] Grabbing container id, name and main VNC port to export it and use it globally.
  [BUA-864] Removing not used code.
  [BUA-864] Making SQ happy.
  [BUA-864] Removing unused imports.
  [BUA-864] Making SQ happy.
  [BUA-864] Updating updating mockito, slf4j, maven compiler plugin
  [BUA-864] Updating docker-client plugin version and fixing unit tests
  ...

# Conflicts:
#	pom.xml
#	src/main/java/de/zalando/tip/zalenium/proxy/DockerSeleniumStarterRemoteProxy.java
#	src/main/java/de/zalando/tip/zalenium/proxy/TestingBotRemoteProxy.java
#	src/test/java/de/zalando/tip/zalenium/proxy/DockerSeleniumRemoteProxyTest.java
#	src/test/java/de/zalando/tip/zalenium/proxy/DockerSeleniumStarterRemoteProxyTest.java
#	src/test/java/de/zalando/tip/zalenium/servlet/LiveNodeServletTest.java
#	src/test/java/de/zalando/tip/zalenium/util/TestUtils.java

+++ scripts/entry.sh
@@ -11,7 +11,21 @@ else
     # Grab the complete docker version `1.12.5` out of the partial one `1.12`
     export DOCKER_VERSION=$(ls /usr/bin/docker-${DOCKER}* | grep -Po '(?<=docker-)([a-z0-9\.]+)' | head -1)
     # Link the docker binary to the selected docker version via e.g. `-e DOCKER=1.11`
-    sudo ln -s /usr/bin/docker-${DOCKER_VERSION} /usr/bin/docker
+    if [ -f /usr/bin/docker-${DOCKER_VERSION} ]; then
+        sudo ln -s /usr/bin/docker-${DOCKER_VERSION} /usr/bin/docker
+    else
+        echo "Something went wrong trying to find DOCKER_VERSION=${DOCKER_VERSION} with DOCKER=${DOCKER}"
+        ls -la /usr/bin/docker*
+        echo "Trying to fetch latest docker binary at /usr/bin/docker*"
+        DOCKER_BIN=$(ls -U /usr/bin/docker* | sort -r | head -1)
+        if [ -f ${DOCKER_BIN} ]; then
+            sudo ln -s ${DOCKER_BIN} /usr/bin/docker
+        else
+            echo "FATAL: Last attempt to find a valid docker binary to use failed."
+            echo "FATAL: DOCKER_BIN=${DOCKER_BIN}"
+            exit 1
+        fi
+    fi
 fi
 
 # Make sure Docker works before continuing
